+++ src/social/components/Comment/index.js
@@ -9,7 +9,6 @@ import customizableComponent from '~/core/hocs/customization';
 import useComment from '~/social/hooks/useComment';
 import CommentComposeBar from '~/social/components/CommentComposeBar';
 import CommentList from '~/social/components/CommentList';
-import ConditionalRender from '~/core/components/ConditionalRender';
 import { notification } from '~/core/components/Notification';
 import { isModerator } from '~/helpers/permissions';
 import StyledComment from './Comment.styles';
@@ -139,7 +138,7 @@ const Comment = ({ readonly = false, commentId, currentUserId, userRoles }) => {
   };
 
   const handleEdit = () => {
-    const { metadata, mentionees } = extractMetadata(markup, mentions);
+    const { metadata, mentionees } = extractMetadata(mentions);
     handleEditComment(localText, mentionees, metadata);
 
     clearAll();
@@ -227,7 +226,7 @@ const Comment = ({ readonly = false, commentId, currentUserId, userRoles }) => {
         isExpanded={isExpanded}
       />
 
-      <ConditionalRender condition={isReplying}>
+      {isReplying && (
         <CommentComposeBar
           postId={comment?.referenceId}
           postType={comment?.referenceType}
@@ -238,7 +237,7 @@ const Comment = ({ readonly = false, commentId, currentUserId, userRoles }) => {
             setExpanded(true);
           }}
         />
-      </ConditionalRender>
+      )}
     </CommentBlock>
   );
 };