+++ src/social/components/CommunityInfo/UICommunityInfo.js
@@ -1,10 +1,10 @@
+import { CommunityPostSettings } from '@amityco/js-sdk';
 import React from 'react';
 import PropTypes from 'prop-types';
 import { toHumanString } from 'human-readable-numbers';
 import Truncate from 'react-truncate-markup';
 import { FormattedMessage, useIntl } from 'react-intl';
 
-import ConditionalRender from '~/core/components/ConditionalRender';
 import customizableComponent from '~/core/hocs/customization';
 import Button from '~/core/components/Button';
 import { PendingPostsBanner } from '~/social/components/CommunityInfo/PendingPostsBanner';
@@ -45,7 +45,7 @@ const UICommunityInfo = ({
   canLeaveCommunity,
   canReviewPosts,
   name,
-  needApprovalOnPostCreation,
+  postSetting,
 }) => {
   const { formatMessage } = useIntl();
 
@@ -80,7 +80,8 @@ const UICommunityInfo = ({
               </div>
             </Count>
           </CountsContainer>
-          <ConditionalRender condition={isJoined}>
+
+          {isJoined && (
             <OptionMenu
               data-qa-anchor="social-community-3dots"
               options={[
@@ -94,7 +95,7 @@ const UICommunityInfo = ({
                 },
               ].filter(Boolean)}
             />
-          </ConditionalRender>
+          )}
         </Header>
 
         {description && (
@@ -103,12 +104,13 @@ const UICommunityInfo = ({
           </Truncate>
         )}
 
-        <ConditionalRender condition={!isJoined}>
+        {!isJoined && (
           <JoinButton onClick={() => joinCommunity(communityId)}>
             <PlusIcon /> <FormattedMessage id="community.join" />
           </JoinButton>
-        </ConditionalRender>
-        <ConditionalRender condition={isJoined && canEditCommunity}>
+        )}
+
+        {isJoined && canEditCommunity && (
           <Button
             fullWidth
             data-qa-anchor="social-edit-community-button"
@@ -116,11 +118,13 @@ const UICommunityInfo = ({
           >
             <PencilIcon /> <FormattedMessage id="community.editProfile" />
           </Button>
-        </ConditionalRender>
-
-        {needApprovalOnPostCreation && isJoined && pendingPostsCount > 0 && (
-          <PendingPostsBanner canReviewPosts={canReviewPosts} postsCount={pendingPostsCount} />
         )}
+
+        {postSetting === CommunityPostSettings.ADMIN_REVIEW_POST_REQUIRED &&
+          isJoined &&
+          pendingPostsCount > 0 && (
+            <PendingPostsBanner canReviewPosts={canReviewPosts} postsCount={pendingPostsCount} />
+          )}
       </Content>
     </Container>
   );
@@ -141,7 +145,7 @@ UICommunityInfo.propTypes = {
   canLeaveCommunity: PropTypes.bool,
   canReviewPosts: PropTypes.bool,
   name: PropTypes.string,
-  needApprovalOnPostCreation: PropTypes.bool,
+  postSetting: PropTypes.string,
   onClickLeaveCommunity: PropTypes.func,
   onEditCommunity: PropTypes.func,
 };