+++ src/social/components/Feed/index.js
@@ -1,6 +1,5 @@
 import React, { memo } from 'react';
 import PropTypes from 'prop-types';
-import { FormattedMessage } from 'react-intl';
 import { PostTargetType, FeedType, CommunityFilter } from '@amityco/js-sdk';
 import DefaultPostRenderer from '~/social/components/post/Post/DefaultPostRenderer';
 
@@ -51,15 +50,10 @@ const Feed = ({
       dataLength={posts.length}
       next={loadMore}
       hasMore={hasMore}
-      loader={
-        <h4>
-          <FormattedMessage id="loading" />
-        </h4>
-      }
     >
       <ConditionalRender condition={!isHiddenProfile}>
         <>
-          <ConditionalRender condition={showPostCreator}>
+          {showPostCreator && (
             <PostCreator
               targetType={targetType}
               targetId={targetId}
@@ -69,7 +63,7 @@ const Feed = ({
               loadMoreCommunities={loadMoreCommunities}
               onCreateSuccess={onPostCreated}
             />
-          </ConditionalRender>
+          )}
 
           {loading && renderLoadingSkeleton()}